redirect back to login page if not logged in

+++ src/routes.py
@@ -19,14 +19,23 @@ def flash_error(error):
 
 @app.route("/")
 def index():
-    if session:
-        if "username" in session.keys():
-            return redirect("/homepage")
+    if not session:
+        return render_template("index.html")
+
+    if not "username" in session.keys():
+        return render_template("index.html")
             
-    return render_template("index.html")
+    return redirect("/homepage")
 
 @app.route("/homepage")
 def homepage():
+
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     communities = communities_service.get_communities(session["username"])
     threads = users_service.get_threads(session["username"])
     return render_template("homepage.html", threads=threads, communities=communities)
@@ -73,6 +82,13 @@ def register():
 
 @app.route("/create_community", methods=["GET", "POST"])
 def create_community():
+
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     if request.method == "GET":
         return render_template("create_community.html")
     if request.method == "POST":
@@ -89,6 +105,13 @@ def create_community():
 
 @app.route("/community/<community_name>")
 def community(community_name):
+
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     community = communities_service.get_community(community_name, session["username"])
     threads = communities_service.get_threads(community_name)
     users = communities_service.get_users(community_name)
@@ -96,11 +119,24 @@ def community(community_name):
 
 @app.route("/join/<community_name>")
 def join(community_name):
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     communities_service.join_community(community_name, session["username"])
     return redirect("/")
 
 @app.route("/create_a_thread", methods=["GET", "POST"])
 def create_a_thread():
+
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     if request.method == "GET":
         communities = communities_service.get_communities(session["username"])
         return render_template("create_a_thread.html", communities=communities)
@@ -121,12 +157,26 @@ def create_a_thread():
 
 @app.route("/thread/<int:thread_id>", methods=["GET", "POST"])
 def thread(thread_id):
+
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     thread = threads_service.get_thread(thread_id, session["username"])
     messages = threads_service.get_messages(thread_id, session["username"])
     return render_template("thread.html", thread=thread, messages=messages)
 
 @app.route("/message/<int:thread_id>", methods=["POST"])
 def message(thread_id):
+
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     content = request.form["content"]
 
     input_validation = validation_service.validate_message(content)
@@ -141,6 +191,13 @@ def message(thread_id):
 
 @app.route("/upvote/<int:thread_id>", methods=["GET"])
 def upvote(thread_id):
+
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     user = users_service.get_user_by_name(session["username"])
     threads_service.upvote(thread_id, user.id)
     flash("Äänestetty", "success")
@@ -148,6 +205,12 @@ def upvote(thread_id):
 
 @app.route("/downvote/<int:thread_id>", methods=["GET"])
 def downvote(thread_id):
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     user = users_service.get_user_by_name(session["username"])
     threads_service.downvote(thread_id, user.id)
     flash("Äänestetty", "success")
@@ -155,6 +218,12 @@ def downvote(thread_id):
 
 @app.route("/upvote/message/<int:thread_id>/<int:message_id>", methods=["GET"])
 def upvote_message(thread_id, message_id):
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     user = users_service.get_user_by_name(session["username"])
     messages_service.upvote(message_id, user.id)
     flash("Äänestetty", "success")
@@ -162,6 +231,12 @@ def upvote_message(thread_id, message_id):
 
 @app.route("/downvote/message/<int:thread_id>/<int:message_id>", methods=["GET"])
 def downvote_message(thread_id, message_id):
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     user = users_service.get_user_by_name(session["username"])
     messages_service.downvote(message_id, user.id)
     flash("Äänestetty", "success")
@@ -169,18 +244,39 @@ def downvote_message(thread_id, message_id):
 
 @app.route("/delete/thread/<int:thread_id>", methods=["GET"])
 def delete_thread(thread_id):
+
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     threads_service.delete_thread(thread_id)
     flash("Ketju poistettu", "success")
     return redirect("/")
 
 @app.route("/delete/message/<int:thread_id>/<int:message_id>", methods=["GET"])
 def delete_message(thread_id, message_id):
+
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     messages_service.delete(message_id)
     flash("Viesti poistettu", "success")
     return redirect(f"/thread/{thread_id}")
 
 @app.route("/edit/thread/<int:thread_id>", methods=["GET", "POST"])
 def edit_thread(thread_id):
+
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     if request.method == "GET":
         thread = threads_service.get_thread(thread_id, session["username"])
         return render_template("edit_thread.html", thread=thread)
@@ -201,6 +297,13 @@ def edit_thread(thread_id):
 
 @app.route("/edit/message/<int:thread_id>/<int:message_id>", methods=["GET", "POST"])
 def edit_message(thread_id, message_id):
+
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     if request.method == "GET":
         message = messages_service.get_message(message_id)
         return render_template("edit_message.html", message=message, thread_id=thread_id)
@@ -212,25 +315,46 @@ def edit_message(thread_id, message_id):
         if not input_validation.passed:
             flash(input_validation.error_message, "warning")
             return redirect(f"/edit/message/{thread_id}/{message_id}")
-            
+
         messages_service.edit(message_id, new_content)
         flash("Muokattu", "success")
         return redirect(f"/thread/{thread_id}")
 
 @app.route("/admin/<community_name>/<username>")
 def admin(community_name, username):
+
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     communities_service.add_admin(community_name, username)
     flash("Lisätty ylläpitäjä", "success")
     return redirect(f"/community/{community_name}")
 
 @app.route("/ban/<community_name>/<username>")
 def ban(community_name, username):
+
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     communities_service.ban(community_name, username)
     flash("Estetty käyttäjä", "success")
     return redirect(f"/community/{community_name}")
 
 @app.route("/leave/<community_name>")
 def leave(community_name):
+
+    if not session:
+        return redirect("/")
+
+    if not "username" in session.keys():
+        return redirect("/")
+
     communities_service.leave(community_name, session["username"])
     flash("Lähdetty ryhmästä", "success")
     return redirect("/")
